Moved hardcoded English string into the i18n file
Tweaked the grammar a bit, removed references to Gabriel Wicke (he wrote the MonoBook skin, but he has had nothing to do with this particular skin AFAIK) and changed "theme" to "skin" because in the MediaWiki world those two words aren't exactly interchangeable.
Quoth self:
"A theme is a set of predefined colors and other such styles for a skin. In essence, a theme can be as simple as one CSS file. Skins on the other hand always have some PHP code and whatnot, in addition to CSS(/LESS) and/or images and/or JavaScript code.
This distinction is important in the MediaWiki world. Other projects written in the PHP programming language, such as phpBB or WordPress, use terms like "templates", "themes" and "skins" interchangeably to refer to various ways to change the look & feel of a website, but those words are not interchangeable in the MediaWiki world! 

See the Theme extension's infopage (https://www.mediawiki.org/wiki/Extension:Theme) on MediaWiki.org for more info."